Install test dagger repository module for uiTests
