+++ app/src/androidTest/java/com/cryart/sabbathschool/ui/splash/SplashActivityTest.kt
@@ -26,35 +26,61 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.Intents.intended
 import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.cryart.sabbathschool.lessons.ui.quarterlies.QuarterliesActivity
 import com.cryart.sabbathschool.ui.login.LoginActivity
+import com.google.firebase.auth.FirebaseAuth
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
+import javax.inject.Inject
+import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.tasks.await
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import org.junit.runner.RunWith
 
 @HiltAndroidTest
+@RunWith(AndroidJUnit4::class)
 class SplashActivityTest {
 
     @get:Rule
     var hiltRule = HiltAndroidRule(this)
 
-    private lateinit var scenario: ActivityScenario<SplashActivity>
+    @Inject
+    lateinit var firebaseAuth: FirebaseAuth
+
+    private var scenario: ActivityScenario<SplashActivity>? = null
 
     @Before
     fun setup() {
+        // Populate @Inject fields in test class
+        hiltRule.inject()
+
+        firebaseAuth.signOut()
+
         Intents.init()
     }
 
     @After
     fun cleanUp() {
-        scenario.close()
+        Intents.release()
+        scenario?.close()
     }
 
     @Test
     fun shouldLaunchLogin() {
         scenario = ActivityScenario.launch(SplashActivity::class.java)
         intended(hasComponent(LoginActivity::class.java.name))
+    }
+
+    @Test
+    fun shouldLaunchQuarterliesActivity() = runBlocking {
+        // Requires signed-in user
+        firebaseAuth.signInAnonymously().await()
+
+        scenario = ActivityScenario.launch(SplashActivity::class.java)
+        intended(hasComponent(QuarterliesActivity::class.java.name))
     }
 }
