+++ app/src/main/java/com/cryart/sabbathschool/ui/login/LoginActivity.kt
@@ -34,7 +34,6 @@ import com.cryart.sabbathschool.core.extensions.arch.observeNonNull
 import com.cryart.sabbathschool.core.model.ViewState
 import com.cryart.sabbathschool.databinding.SsLoginActivityBinding
 import com.cryart.sabbathschool.databinding.SsLoginButtonsBinding
-import com.cryart.sabbathschool.lessons.ui.lessons.SSLessonsActivity
 import com.cryart.sabbathschool.lessons.ui.quarterlies.QuarterliesActivity
 import com.facebook.CallbackManager
 import com.facebook.login.LoginManager
@@ -134,18 +133,6 @@ class LoginActivity : AppCompatActivity() {
     private fun launchMain() {
         startActivity(
             Intent(this, QuarterliesActivity::class.java).apply {
-                flags = Intent.FLAG_ACTIVITY_CLEAR_TOP
-            }
-        )
-        finish()
-    }
-
-    /**
-     *  Open Lessons Activity from Features / Lessons
-     */
-    private fun launchLessons() {
-        startActivity(
-            Intent(this, SSLessonsActivity::class.java).apply {
                 flags = Intent.FLAG_ACTIVITY_CLEAR_TOP
             }
         )
